Separate IslandName argument from PlayerName argument
Currently, the first argument is used both as the island name and the
player name.

This makes usage of the command substantially less flexible, and
prevents one of the old use cases for this command (players moving
themselves to a group island).

This commit separates out the two arguments to allow for usage with e.g.
@p in command blocks, but still allow specification of the destination
island independent of the player name.